Update ingress creation menu
