+++ pkg/model/ingress/activeingress/edit.go
@@ -6,7 +6,9 @@ import (
 	"strings"
 
 	"github.com/containerum/chkit/pkg/model/ingress"
+	"github.com/containerum/chkit/pkg/porta"
 	"github.com/containerum/chkit/pkg/util/activekit"
+	"github.com/containerum/chkit/pkg/util/ferr"
 	"github.com/containerum/chkit/pkg/util/text"
 	"github.com/containerum/chkit/pkg/util/tlsview"
 	"github.com/sirupsen/logrus"
@@ -54,6 +56,19 @@ func EditWizard(config Config) (ingress.Ingress, error) {
 						}
 						border := strings.Repeat("_", text.Width(data))
 						fmt.Printf("%s\n%s\n%s\n", border, data, border)
+						return nil
+					},
+				},
+				{
+					Label: "Export ingress to file",
+					Action: func() error {
+						var fname = activekit.Promt("Type filename: ")
+						fname = strings.TrimSpace(fname)
+						if fname != "" {
+							if err := (porta.Exporter{OutFile: fname}.Export(ingr)); err != nil {
+								ferr.Printf("unable to export ingress:\n%v\n", err)
+							}
+						}
 						return nil
 					},
 				},
